chore(backend): db init with mikroORM lib
