+++ apps/backend/src/app/app.module.ts
@@ -1,18 +1,20 @@
 import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { DataSource } from 'typeorm';
+import { MikroOrmModule } from '@mikro-orm/nestjs';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { configService } from '../config/config.service';
+//import { configService } from '../config/config.service';
 import { ProjectModule } from '../project/project.module';
-import { UserModule } from '../user/user.module';
-import { AuthModule } from '../auth/auth.module';
+import config from '../config/mikro-orm';
+
 
 @Module({
-  imports: [TypeOrmModule.forRoot(configService.getTypeOrmConfig()), ProjectModule, UserModule, AuthModule],
+  imports: [
+    MikroOrmModule.forRoot({
+      ...config,
+      autoLoadEntities: true,
+    }),
+    ProjectModule],
   controllers: [AppController],
   providers: [AppService],
 })
-export class AppModule {
-  constructor(private dataSource: DataSource) {}
-}
+export class AppModule { }
