+++ apps/backend/src/main.ts
@@ -6,6 +6,7 @@
 import { Logger } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger';
+import { MikroORM } from '@mikro-orm/core';
 
 import { AppModule } from './app/app.module';
 
@@ -20,6 +21,9 @@ async function bootstrap() {
 
   const document = SwaggerModule.createDocument(app, config);
   SwaggerModule.setup('api', app, document);
+
+  await app.get(MikroORM).getSchemaGenerator().ensureDatabase();
+  await app.get(MikroORM).getSchemaGenerator().updateSchema();
 
   const globalPrefix = 'api';
   app.setGlobalPrefix(globalPrefix);
