+++ dev.docker-compose.yml
@@ -6,9 +6,17 @@ services:
     image: kartoza/postgis:15-3.3
     restart: always
     environment:
-      - POSTGRES_DB=gis
       - POSTGRES_USER=docker
       - POSTGRES_PASS=docker
       - ALLOW_IP_RANGE=0.0.0.0/0
     ports:
-      - '5432:5432'
+      - 5432:5432
+  pgadmin:
+    container_name: pgadmin4_container
+    image: dpage/pgadmin4
+    restart: always
+    environment:
+      PGADMIN_DEFAULT_EMAIL: admin@admin.com
+      PGADMIN_DEFAULT_PASSWORD: root
+    ports:
+      - "5431:80"
