+++ package.json
@@ -35,7 +35,7 @@
     "@types/jest": "^28.1.1",
     "@types/node": "18.11.9",
     "@types/passport-jwt": "^3.0.8",
-    "@types/passport-local": "^1.0.35",
+    "@types/passport-local": "^1.0.34",
     "@types/react": "18.0.25",
     "@types/react-dom": "18.0.9",
     "@types/react-is": "17.0.3",
@@ -86,25 +86,19 @@
     "@docusaurus/theme-mermaid": "^2.2.0",
     "@mdx-js/react": "^1.6.22",
     "@mikro-orm/core": "^5.6.8",
+    "@mikro-orm/nestjs": "^5.1.6",
     "@mikro-orm/postgresql": "^5.6.8",
     "@nestjs/cli": "^9.1.8",
     "@nestjs/common": "^9.0.0",
     "@nestjs/core": "^9.0.0",
-    "@nestjs/jwt": "^10.0.1",
-    "@nestjs/passport": "^9.0.0",
     "@nestjs/platform-express": "^9.0.0",
     "@nestjs/swagger": "^6.1.4",
-    "@nestjs/typeorm": "^9.0.1",
     "@reduxjs/toolkit": "^1.9.1",
     "assert": "^2.0.0",
     "babel-plugin-formatjs": "^10.3.35",
     "core-js": "^3.6.5",
-    "crypto": "^1.0.1",
     "kepler.gl": "^2.5.5",
     "msw": "^0.49.2",
-    "passport": "^0.6.0",
-    "passport-jwt": "^4.0.1",
-    "passport-local": "^1.0.0",
     "pg": "^8.8.0",
     "prism-react-renderer": "^1.3.5",
     "prop-types": "^15.8.1",
@@ -123,8 +117,7 @@
     "regenerator-runtime": "0.13.7",
     "rxjs": "^7.0.0",
     "styled-components": "5.3.6",
-    "tslib": "^2.3.0",
-    "typeorm": "^0.3.11"
+    "tslib": "^2.3.0"
   },
   "browserslist": {
     "production": [
